feat: Widen Tick to u64; from_ticks returns Option<Time>
The master `Tick` counter widens from `u32` to `u64`. With the wider
counter, `time_to_tick`'s `beats: u32 × tick_count: u32` arithmetic
fits in `u64` with ~32 bits of headroom — the `checked_mul().expect()`
backstop is no longer needed. `arb_time` is freed to use
`any::<u32>()` for `beats` (full type domain) instead of the
clamped `0..=100_000` that masked the overflow.

`from_ticks` and `from_ticks_floor` now return `Option<Time>` because
`Time.beats` stays `u32`: `Tick(huge)` values where `huge / chosen_tc`
exceeds `u32::MAX` have no representable `Time`. The `ticks` Conn
unwraps under a documented precondition (`arb_tick` capped at
`u32::MAX × Grid::T1.tick_count()`); runtime callers handle `None`
themselves.

Cast updates ripple through `swing.rs` (i64 → i128 widening),
`sample_tick.rs` (u128 saturation cap → u64::MAX), `transform.rs`,
`scheduler.rs`, `envelope.rs` (linear_u8/smoothstep_u8 widened to
u64), and `cli` test fixtures.

+++ crates/cli/src/channel_trace.rs
@@ -30,7 +30,7 @@ pub struct TraceArgs {
 pub struct TraceRow {
     pub buffer_index: u32,
     pub sample_index: u64,
-    pub tick: u32,
+    pub tick: u64,
 }
 
 /// Pure CPU scheduling trace — useful for testing without capturing
